Revert "[MDS-5196] - Common package Typescript setup (#2510)"
This reverts commit d82a1a58b2ee84d8c3a47586d303a6deb3822daf.
