+++ .eslintrc.js
@@ -1,34 +1,16 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
 const path = require("path");
 
 module.exports = {
-  root: true,
-  parser: "@typescript-eslint/parser",
-  parserOptions: {
-    project: [
-      "./services/common/tsconfig.json",
-      "./services/minespace-web/tsconfig.json",
-      "./services/core-web/tsconfig.json",
-    ],
-  },
+  parser: "babel-eslint",
   env: {
     browser: true,
     node: true,
     "jest/globals": true,
     es6: true,
   },
-  plugins: ["jest", "@typescript-eslint"],
-  extends: [
-    "eslint:recommended",
-    "plugin:react/recommended",
-    "plugin:@typescript-eslint/recommended",
-    "airbnb-typescript",
-    "prettier",
-  ],
+  plugins: ["jest"],
+  extends: ["eslint:recommended", "plugin:react/recommended", "airbnb", "prettier"],
   settings: {
-    "import/parser": {
-      "typescript-eslint-parser": [".ts", ".tsx"],
-    },
     "import/resolver": {
       alias: {
         map: [
@@ -56,17 +38,5 @@ module.exports = {
     "jsx-a11y/label-has-for": 0, // deprecated rule
     "import/no-cycle": 0,
     "no-param-reassign": 0,
-    "no-shadow": 0,
-    "@typescript-eslint/no-shadow": 1,
-    "@typescript-eslint/no-unused-vars": 1,
-    "import/no-extraneous-dependencies": 0,
-    "@typescript-eslint/quotes": [2, "double"],
-    "@typescript-eslint/indent": [0, 4],
-    "@typescript-eslint/comma-dangle": 0,
-    "@typescript-eslint/naming-convention": 0,
-    "react/prop-types": 0,
-    "@typescript-eslint/ban-ts-comment": 0,
-    "@typescript-eslint/default-param-last": 0,
-    "@typescript-eslint/no-explicit-any": 1,
   },
 };
