+++ package.json
@@ -14,11 +14,8 @@
     ]
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.59.1",
-    "@typescript-eslint/parser": "^5.59.1",
     "eslint": "7.27.0",
     "eslint-config-airbnb": "18.2.1",
-    "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "6.9.0",
     "eslint-import-resolver-alias": "1.1.2",
     "eslint-plugin-import": "2.26.0",
@@ -38,7 +35,7 @@
       "prettier --write --ignore-unknown --config .prettierrc.json"
     ],
     "*.{js,ts,tsx}": [
-      "eslint --fix --max-warnings=20"
+      "eslint --fix"
     ]
   },
   "lint-prepush": {
