+++ services/common/package.json
@@ -7,15 +7,14 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "build": "webpack --env production && tsc --declaration --emitDeclarationOnly",
-    "watch": "webpack --watch --watch-poll --hot --env development & tsc --watch --declaration --emitDeclarationOnly"
+    "build": "webpack --env production",
+    "watch": "webpack --watch --watch-poll --hot --env development"
   },
   "dependencies": {
     "@babel/runtime": "7.15.3",
     "@babel/runtime-corejs3": "7.15.3",
     "query-string": "5.1.1",
-    "react": "16.13.1",
-    "ts-loader": "8.4.0"
+    "react": "16.13.1"
   },
   "devDependencies": {
     "@babel/core": "7.15.0",
@@ -39,7 +38,6 @@
     "react-hot-loader": "4.12.21",
     "style-loader": "0.23.1",
     "terser-webpack-plugin": "1.3.0",
-    "typescript": "^5.0.4",
     "webpack": "4.41.5",
     "webpack-cli": "3.3.10",
     "webpack-dev-server": "3.11.0",
