--- services/common/src/constants/enums.ts
@@ -1,14 +1,0 @@
-export enum NoDTypeEnum {
-  non_substantial = "Non Substantial",
-  potentially_substantial = "Potentially Substantial",
-}
-
-export enum NoDStatusEnum {
-  pending_review = "Pending Review",
-  in_review = "In Review",
-  information_required = "Information Required",
-  self_determined_non_substantial = "Self Determined Non Substantial",
-  determined_substantial = "Determined Substantial",
-  determined_non_substantial = "Determined Non Substantial",
-  withdrawn = "Withdrawn",
-}
