+++ services/common/src/constants/index.js
@@ -7,13 +7,8 @@
  */
 export * from "./actionTypes";
 export * from "./API";
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
 export * from "./reducerTypes";
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-export * from "./strings";
 export * from "./environment";
+export * from "./strings";
 export * from "./fileTypes";
 export * from "./utils";
-export * from "./enums";
