--- services/common/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./constants";
-export * from "./interfaces";
