--- services/common/src/interfaces/amendment.interface.ts
@@ -1,9 +1,0 @@
-export interface IAmendment {
-  permit_amendment_guid: string;
-  permit_amendment_status_code: string;
-  permit_amendment_type_code: string;
-  permit_ammendment_id: string;
-  received_date: string;
-  issue_date: string;
-  authorization_end_date: string;
-}
