--- services/common/src/interfaces/index.ts
@@ -1,5 +1,0 @@
-export * from "./noticeOfDeparture.interface";
-export * from "./permit.interface";
-export * from "./mine.interface";
-export * from "./tailingsStorageFacility.interface";
-export * from "./amendment.interface";
