--- services/common/src/interfaces/tailingsStorageFacility.interface.ts
@@ -1,5 +1,0 @@
-export interface ITailingsStorageFacility {
-  mine_guid: string;
-  mine_tailings_storage_facility_guid: string;
-  mine_tailings_storage_facility_name: string;
-}
