+++ services/core-web/common/selectors/noticeOfDepartureSelectors.js
@@ -1,3 +1,4 @@
 import * as noticesOfDepartureReducer from "../reducers/noticeOfDepartureReducer";
 
+// eslint-disable-next-line import/prefer-default-export
 export const { getNoticesOfDeparture, getNoticeOfDeparture } = noticesOfDepartureReducer;
