+++ services/minespace-web/src/App.js
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { compose, bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
@@ -23,9 +23,6 @@ import configureStore from "./store/configureStore";
 
 export const store = configureStore();
 
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
-
 Spin.setDefaultIndicator(<LoadingOutlined style={{ fontSize: 40 }} />);
 
 const propTypes = {
@@ -39,13 +36,7 @@ const defaultProps = {
   staticContentLoadingIsComplete: false,
 };
 
-interface AppProps {
-  isAuthenticated: boolean;
-  staticContentLoadingIsComplete: boolean;
-  loadBulkStaticContent: () => void;
-}
-
-const App: FC<AppProps> = (props) => {
+const App = (props) => {
   const [isIE, setIsIE] = useState(true);
   const [isMobile, setIsMobile] = useState(true);
 
@@ -55,7 +46,7 @@ const App: FC<AppProps> = (props) => {
     if (isAuthenticated) {
       loadBulkStaticContent();
     }
-    setIsIE(!!detectIE());
+    setIsIE(detectIE());
   }, []);
 
   useEffect(() => {
@@ -87,7 +78,7 @@ const App: FC<AppProps> = (props) => {
               <MediaQuery maxWidth={500}>
                 {isMobile && <WarningBanner type="mobile" onClose={handleMobileWarningClose} />}
               </MediaQuery>
-              <Row justify="center" align="top">
+              <Row type="flex" justify="center" align="top">
                 <Col xs={xs} lg={lg} xl={xl} xxl={xxl}>
                   <Routes />
                 </Col>
@@ -124,5 +115,4 @@ export default compose(
   hot(module),
   connect(mapStateToProps, mapDispatchToProps),
   AuthenticationGuard(true)
-  // @ts-ignore
 )(App);
